experimenting with writing in markdown, including pics
I managed to make a markdown file and get it working, but really, have i written anything to show that i can be flexible with this platform? I'm uploading these images to use in a small document. not a story, just talking about the images, maybe giving some context to them. It'll probably be more like a journal entry than a story, but i feel like this is a needed step above what i had done so far.